♻️ FastAPIのデータ型をschemasファイルに分割
